airflow/requirements.txt
psycopg-binary
jupyter
jupyter-core
airflow-dbt
apache-airflow-providers-jdbc
airflow-providers-clickhouse
apache-airflow-providers-amazon

Faker
psycopg
python-dateutil
python-dotenv
six
typing_extensions
ruff
agate
annotated-types
attrs
babel
certifi
charset-normalizer
click
colorama
daff
dbt-adapters
dbt-common
dbt-core
dbt-extractor
dbt-semantic-interfaces
dbt-starrocks
deepdiff
Faker
idna
importlib-metadata
isodate
Jinja2
jsonschema
jsonschema-specifications
leather
Logbook
MarkupSafe
mashumaro
minimal-snowplow-tracker
more-itertools
msgpack
mysql-connector-python
networkx
ordered-set
packaging
parsedatetime
pathspec
protobuf
pydantic
pydantic_core
python-dateutil
python-dotenv
python-slugify
pytimeparse
pytz
PyYAML
referencing
requests
rpds-py
ruff
six
sqlparse
text-unidecode
typing_extensions
urllib3
zipp
PyMySQL
starrocks

# Зависимости для Airflow с поддержкой кастомного оператора S3XlsxToStarRocksOperator

# Основные зависимости для работы с Excel
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.0

# Зависимости для работы с S3/MinIO
boto3>=1.26.0
botocore>=1.29.0

# Зависимости для работы с MySQL/StarRocks
pymysql>=1.0.0
sqlalchemy>=1.4.0

# Дополнительные зависимости для Airflow
apache-airflow-providers-amazon>=8.0.0
apache-airflow-providers-mysql>=4.0.0

apache-airflow-providers-cncf-kubernetes==10.8.2

astronomer-cosmos==1.11.0

kafka-python